Make get_state independent of exclude list
Getting the state of a device now depends only on the default_state
field which simplifies things. It also fixes a bug - if a device was in
the exclude list and had a default_state field, the latter was ignored
which could potentially result in the wrong state being displayed.